removing specifc impl of cartesian plot
